+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog for DP3T-SDK iOS
 
+## Latest
+- Fixed issue contact date calculation
+
 ## Version 0.1.8 (27.4.2020)
 - Add parameter to set bucket length (has to be supported by the backend)
 - Sets reconnectionDelay to 1 minute